1차 빌드

+++ Assets/Scripts/ProjectTile.cs
@@ -6,10 +6,10 @@ public class ProjectTile : MonoBehaviour
 {
     private Movement2D movement2D;
     private Transform target;
-    private int damage;
+    private float damage;
 
     
-    public void Setup(Transform target, int damage)
+    public void Setup(Transform target, float damage)
     {
         movement2D = GetComponent<Movement2D>();
         this.target = target;
